oauth-proxy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
redis==5.0.1
slowapi==0.1.9
pydantic==2.5.0
python-multipart==0.0.6